Intro Screen for Earth Entry
I worked on the GUI layout for the Earth's first mission selector and I intent to add a short description of Earth and let the player choose the first level as the others will be locked. Also, worked on a pause menu that is also activates with the "escape" button.