I need a redbull...
Essentially, I've integrated Main Menu, Earth Mission Select Scene and Level 1 with ambience music. I touched a bit in the animations such as making 4 animations where the player tilts right, up, down and left. Also, I've successfully finished the settings tab where it will be probably renamed to "How to Play" unless I decide to figure out if I really want to make sliders for the music volume and things like that! Enemies now have a short delay before they attack the user and I'll probably implement a bit of text at the top to tell the player that they should avoid their bullets. Secondly, I added arrow keys into the GUI so the player at least has some little knowledge on what keys to use in order to move around so that they don't have to remember what the controls are all the way at the main menu, as well as other buttons such as the rockets button/ammo so that at least the player knows what they can and can't do in game. In addition, I've added sound effects/VFX to the UI buttons for a more realistic SCI-FI feel and plan on troubleshooting sound effects for the rings on level 1 plus enemy/player bullets and hit markers. I plan to add VFX for when the enemy has been hit but not fully killed by the player as well as importing a new particle asset to make the fire a bit more realistic.

I gave up on the enemies moving for the time being plus the ship tilt but I know I can target those problems after I mold and settle some core gameplay. Things such as cutscenes and such will be handled by joshua as well as the character portrait. I'll probably include things such as AI voice, VFX for the dialogue and some type of animation so that the player can read as "Captain J" types them instructions from HQ.

Some more nitpicky details would be that I fixed the lighting in main menu so now the player can see certain aspects of the ship and the ship itself. I want to add some sort of animation or blinking lights like a faulty lamppost but that would be for tomorrow me.
I'm sorry Mr. Shams I did what I could ;-;